refactor: rename HashAlgorithm to HashTy and use uint8_t enums
Renamed HashAlgorithm enum to HashTy across the codebase and changed underlying type to uint8_t for memory efficiency. Updated all references, constructors, and method signatures accordingly. Also standardized enum underlying types in other enums (GcmCryptoSchema, BlockCipherTy, PRG_MODE, etc.) to uint8_t and fixed header include order in several files.

+++ yacl/aead/all_gcm.h
@@ -14,9 +14,8 @@
 
 #pragma once
 
-#include <vector>
-
 #include <span>
+#include <vector>
 
 #include "yacl/utils/byte_container_view.h"
 #include "yacl/utils/secparam.h"
@@ -41,7 +40,7 @@ constexpr size_t kSm4GcmMacSize = 16;
 constexpr size_t kSm4GcmKeySize = 16;
 #endif
 
-enum class GcmCryptoSchema : int {
+enum class GcmCryptoSchema : uint8_t {
   AES128_GCM, /* security level = 128 */
   AES256_GCM, /* security level = 256 */
 #ifdef YACL_WITH_TONGSUO
